Changed myth_system to use a reaper thread.  This takes care of the possibility of several concurrent children exiting at the same time.
The MythSystemReaper thread is started the first time is is to be used, and
uses a QMap of PID -> { mutex, return value } to match up the dying child 
(from waitpid(-1, &status, WNOHANG)) to the appropriate mutex which is blocking
the calling thread.  There is a 100ms timeout between calls to waitpid, causing
the blocking to be in the timeout, not in waitpid() itself as an old workaround
indicated that OSX has potential issues with blocking in waitpid.

Also, in the child process, after the fork, clear the verbose mutex if it is 
locked, as this will cause the child process to freeze should it try to use
VERBOSE.


This *should* fix the bug where starting a recording gets stuck at "Please 
Wait..." waiting for the pulseaudio disabling to return.  myth_system() is
now thread-safe.



git-svn-id: http://svn.mythtv.org/svn/trunk@25977 7dbf422c-18fa-0310-86e9-fd20926502f2