Modify RemoteFile::Read() so that it uses a graduated scale for it's waitForRead select timeout rather than starting at 200ms.
With a local backend, the response from the server can come back
in under 1ms, before the select in waitForRead has even had a
chance to fire off.  This means that ::Read() would always wait
200ms for data that was already there.  This patch slowly ramps
up the timeout starting at 10ms, incrementing by 20ms each
time.  This should decrease the latency in RemoteFile ::Read()
calls, since it takes us almost a full second before we get
around the 200ms timeout range the previous code started with.

It's not the normal case, but in testing the remote Blu-ray
streaming, this patch took my mythavtest startup time from ~30
seconds down to around 5-6.  This was streaming from a BE to
a FE on the same box.  True remote streaming from a remote
host will not see as big of a benefit but there should stil
be some.



git-svn-id: http://svn.mythtv.org/svn/trunk@25987 7dbf422c-18fa-0310-86e9-fd20926502f2