Add a new 'Theme Chooser' screen which can be used to browse and choose a new theme.  This includes local themes by default. From the popup menu, the user has the option of also displaying the list of available downloadable themes.
The user can browse themes using the default buttonlist and
small preview interface, or switch to a full screen preview
mode where a high-resolution full-screen preview of the theme
is shown to see what it would look like on the user's display.

When selecting a new theme, if the theme is local, it is
immediately loaded.  If the theme is remotely downloadble,
the theme is downloaded and installed in the normal config
dir under the 'themes' subdir (ie, ~/.mythtv/themes/LCARS for
the LCARS theme under Linux).

NOTE: The menu option for the new screen is under settings, but
is currently disabled until the theme packages have been pushed
out to the FTP mirrors.  Once this happens during it's regularly
scheduled nightly task (so I can verify it works), then I'll
uncomment the menu item and add menu entries to the menus other
than defaultmenu.

This window is based on code originally by Robert McNamara and he
plans on updating my attempt at default and default-wide screens
to make them more appealing than my simple versions.



git-svn-id: http://svn.mythtv.org/svn/trunk@26002 7dbf422c-18fa-0310-86e9-fd20926502f2