libmythmetadata:  Start moving towards being able to offer a few different scan behaviors.  First, add a convenience method that will simply scan all video directories, hosts, and video storage directories, doScanAll().
Also, remove some of the legacy non-MythUI removal prompting that has been commented out since the MythUI port, and rename the removal method to align the language with recordings-style terminology.

Next steps will be to add flags to start the scan with an explicit option not to delete any orphaned item, but simply to flag it as "missing."  Then we can add some UI to notify the user that items were missing, and get their explicit go-ahead to remove any missing items (or alternately, to go investigate why they are not being found).  At this point, the scan can be fairly easily put into the backend and run on a schedule, as we won't have to worry about metadata being cleared when someone's network card goes down.


git-svn-id: http://svn.mythtv.org/svn/trunk@26032 7dbf422c-18fa-0310-86e9-fd20926502f2