Adds support for children running in the background in myth_system.  If the MYTH_SYSTEM_RUN_BACKGROUND flag is included, or the command ends with &, then the caller will not block waiting for the return anymore.  The child pid is still handed to the reaper so the child doesn't get zombied, but the caller carries on with minimal delay.
This is hopefully a fix for and refers to #8804.  I couldn't effectively test
with a real screensaver call as I don't have them installed, but I did watch
the processes try (after locally pretending to have the commands required).

Also changed the other known commands ending in & to set the
MYTH_SYSTEM_RUN_BACKGROUND flag.



git-svn-id: http://svn.mythtv.org/svn/trunk@26092 7dbf422c-18fa-0310-86e9-fd20926502f2