Fixes #7195. This speeds up preview generation and transport.
It's been some time since preview generation had any real love, it's been updated to deal with storage groups and it's been moved from the frontend to the backend and from an internal thread to an external process and generally a lot of cruft has built up. This dumps a lot of that. It abstracts the preview generation queue that was being used on the frontend so it can run on both the front and back ends and consequently limit the thread count on the backend to the number of cores *2 instead of the number of cores on all frontends * 2 as was effectively the case after it was moved out of the frontend.

There is still at least one major optimization that can be made, documented in the ticket. But this gets the performance up to an acceptable level and is all I have ready before the feature freeze anyway.

Thanks go out to Stuart Morgan who very quickly came up with a fix for a wexing problem in MythUIButtonList which allows preview generation to run well for those using the Arclight and Terra themes (I've had a patch that worked well for the MythCenter themes for some time now :)

Note: distclean is recommended and all clients need to be updated due to a new mythproto version.

git-svn-id: http://svn.mythtv.org/svn/trunk@26106 7dbf422c-18fa-0310-86e9-fd20926502f2