Refactor the user interface of the recording cut list editor to take advantage of the new cleaner design of the editor provided by Mark's libmythui-osd code, which ensures users can't create invalid cut lists with discontinuities.  Refs #7916.  Fixes #8564.  Fixes #8832.
Includes some changes to key bindings to make key usage more consistent with the rest of MythTV.
{{{
 * MENU no longer exits the editor.  It now brings up the menu.
 * SELECT no longer brings up the menu.  It now adds a new cut or, when in a cut area, prompts whether to move the cut start/end or delete the cut.
 * EDIT exits the editor and saves the cut list.
 * ESCAPE prompts whether to save the cut list or cancel changes, then exits.
 * DELETE can be used to delete a cut, when inside a cut area.
 * SAVEMAP (new binding, no default) saves the cut list without exiting (works well when mapped to the same key as TV Editing/TOGGLERECORD).
}}}

The menu allows moving the previous/next cut point to the current position or, if there is no previous/next cut point, cutting to the beginning or end of the recording, and allows adding a new cut.  It also includes an option, "Cut List Options", which brings up a secondary menu allowing the user to "Clear Cut List", "Invert Cut List", "Undo Changes" (revert to the saved cut list), "Exit Without Saving", "Save Cut List" (without exiting), or "Save Cut List and Exit".

A typical editing session would involve:
{{{
 * Hit EDIT (E) to enter edit mode
 * Optionally import the flag list with LOADCOMMSKIP (Z)
 * Find the start (or end) frame for a new cut and press SELECT (Space/Enter)
 * Find the end (or start) frame for that cut and press SELECT (Space/Enter)
 * To adjust positions of existing cuts, find the desired frame, then press MENU (M), then select "Move Previous Cut End Here" or "Move Next Cut Start Here"
 * To remove a cut, press DELETE (D) inside the cut region
 * Hit EDIT (E) to exit edit mode and save the cut list
}}}
All actions could instead be performed with the MENU (M).

Thanks to Jim Stichnoth for the MythUIEditBar fixes and for the patches on #8564 that people have been using while I've been working on this patch.


git-svn-id: http://svn.mythtv.org/svn/trunk@26111 7dbf422c-18fa-0310-86e9-fd20926502f2