This rather large commit removes the unintentional circular dependency we managed to make when moving myth_system to libmythdb in [24623].
Replace the calls to the lirc and joystick menu locks with events to the
main window.  This collapses the flags to myth_system some, which touches a
lot of consumers to use the new flags.  Changed the flags to an enum (at the 
suggestion of cpinkham).

Also removes the classes and events used to do those locks as they are no
longer used anywhere.

Finally, make the blocked threads timeout so they can processEvents every 100ms,
then reblock.  This allows the LIRC presses, etc during that time to get
ignored properly and in a relatively timely manner.

Fixes #8863



git-svn-id: http://svn.mythtv.org/svn/trunk@26166 7dbf422c-18fa-0310-86e9-fd20926502f2