Channel Scanner:  Fix for QAM/SCTE/OpenCable scanning.  A full scan of Cable in the US/Canada should now produce expected channel numbering on a clean database without conflicts.
Previously, the subchannel was being used as the uniqueness identifier.  So by definition, 53-1 and 54-1 and 55-1 were all in conflict, as the subchannel was all the info it had to compare.

The fix is twofold:  First, switch to using channum as the "uniqueness" identifier for MPEG, SCTE, and OpenCable channels.  Unfortunately, these channels weren't being assigned a channum in the channel scanner itself, so when one of these types is present, generate a channum based on frequency identifier (what most US folks consider "physical" channel) and serviceid (what most of us consider "subchannel).  This basically produces channels that are consistent with how every television in the relevant markets works, and should not hurt the scanning/channel naming conventions of any other locale at the same time.  Since MPEG channels are the lowest common denominator and mostly only present on engineering/testing/non-full-fledged channels elsewhere, this channel naming scheme appears safe.

Fixes #7838.  Almost certainly fixes #8078 (that, or one of the other fixes from last night should have).  Refs #7284 (needs a retest after the IsConflicting fixes last night).


git-svn-id: http://svn.mythtv.org/svn/trunk@26177 7dbf422c-18fa-0310-86e9-fd20926502f2