Remove the "back" option from the second menu in the cut list editor.
It was added in response to a misinterpretation of a suggestion from a tester.  Without the menu entry, the menu size is more manageable and usable, and the menu is more consistent with other areas using pop-up style menus.

Note, also, that I may change the edit menu to be a full-fledged OSD menu after the feature freeze, which will obviate the question of whether to have a recursive menu.


git-svn-id: http://svn.mythtv.org/svn/trunk@26219 7dbf422c-18fa-0310-86e9-fd20926502f2