Make MSqlQuery.exec and .prepare more resiliant to disappearing MySQL connections.
If the database connection is down on entry, it will try to 
reconnect or fail.  

If the .exec or .prepare call fails due to "MySQL server went away" (error code
2006), it will try to reconnect, and if that works, it will retry the call once.

This should keep the frontend and backend (et al) from going absolutely crazy
if the mysql server needs to be restarted for any reason.  Previously, all
further queries would fail, requiring restarts.

Fixes #8903



git-svn-id: http://svn.mythtv.org/svn/trunk@26225 7dbf422c-18fa-0310-86e9-fd20926502f2