Disable OpenGL vsync.
This has not worked since the OSD merge and despite trying a number of
different approaches, I've utterly failed to fix it. I suspect it is due
to some interaction with the Qt rendering subsystem, though that doesn't
explain why it would fail when using the Qt painter as well as the
OpenGL version.

This will have no impact on windows or OS X builds (for which it is
unsupported and which manage vsync without it).

On Linux, VDPAU playback will be unaffected. For XVideo, XvMC and OpenGL
you will need to ensure the appropriate sync to vblank settings are
enabled. This is easily done in nvidia-settings for NVidia drivers (and
from memory in a similar way for ATI). Intel generally supports DRM sync
in its newer drivers and also enables XV_SYNC_TO_VBLANK by default for
its XVideo ports.

Going forward, we should probably remove the OpenGL vsync code.


git-svn-id: http://svn.mythtv.org/svn/trunk@26314 7dbf422c-18fa-0310-86e9-fd20926502f2