Fix the channel change timeout popup.
I suspect this was broken by r25543 (Signal monitor changes) and if so
there are sections of code that need to be stripped out as the behaviour
now appears to be triggered by events from the signal monitor and not by
the timer directly.

Closes #8906.

As an aside, the message or the themes probably need to be adjusted as
presentation is generally a mess when using XVideo.


git-svn-id: http://svn.mythtv.org/svn/trunk@26327 7dbf422c-18fa-0310-86e9-fd20926502f2