Open an official, blessed version of the loophole that was closed in [26352].  This fixes the issue with the Theme Chooser screen not refreshing and getting stuck in an endless 'busy dialog' when toggling the 'show downloadable themes' button.
ReloadInBackground() can be used to trigger a background call to
Load() and subsequent call to Init().  This can be used in a
screen to refresh the data if the screen is designed with this in
mind.  The Theme Chooser screen is setup this way.

NOTE: This does bump the binary API version, so make clean, etc..



git-svn-id: http://svn.mythtv.org/svn/trunk@26367 7dbf422c-18fa-0310-86e9-fd20926502f2