Change SSDP discovery to not use HttpServer
There is no need for an HTTP server to do SSDP discovery to find backends.
This changeset removes that from MythContextPrivate::InitUPnP(), and is tested
to work, and quite quickly in my setup (less than half a second).

Instead of using port 6549 as the local end of the SSDP search connection, we
now use 0, causing the port to be automatically selected, and will keep
multiple simultaneous SSDP searching processes from conflicting with each other
trying to bind to a single UDP port.

Should fix the hanging mythpreviewgen processes reported on the -users mailing
list, and quite likely other similar issues (like mythcommflag processes giving
the same odd behavior).  Note:  this SSDP search only seems to be effective if
there is a config.xml file with a USN in the DefaultBackend section.



git-svn-id: http://svn.mythtv.org/svn/trunk@26382 7dbf422c-18fa-0310-86e9-fd20926502f2