Fix an issue that causes the pre-theme locale and/or backend selection prompt to appear full size, but with unreadably tiny fonts in the upper left corner of the screen.
This issue seemed to exist only for users without a database (such as new users without a config.xml/mysql.txt and for whom the database defaults don't work).  The most straight-forward fix was to backport [25410] from trunk.

Thanks to blizzard` and lenswipe|ill on IRC for testing to help confirm the configuration that causes the issue.


git-svn-id: http://svn.mythtv.org/svn/branches/release-0-23-fixes@26394 7dbf422c-18fa-0310-86e9-fd20926502f2