Effectively disable video rendering when embedded and the video is fully obscured. Closes #8520
If we are 'embedding' but don't actually have a preview window, tell the
player to display in an empty rect, this disabling any actual
framebuffer updates but maintaining all other aspects of playback - such
as audio and a/v sync.

This doesn't actually address the root cause of the problem noted in
#8520 (which I think is related to the painter window mask when using
the Qt UI painter) but it seems to be an effective solution without any
unwanted side effects. After all of the recent fixes to embedded
playback and colourkey repainting etc, I could replicate the problem in
#8520 with any video renderer in conjunction with the Qt UI painter.


git-svn-id: http://svn.mythtv.org/svn/trunk@26421 7dbf422c-18fa-0310-86e9-fd20926502f2