Add some additional padding around text based subtitles.
Even when working as expected, the current code has little or no padding
around rendered text - which is noticeable at the sides when using a
black background and top and bottom with certain fonts. For some reason
fonts are also sometimes cropped when the text extends beyond the
calculated width.

So add a small amount of padding around each line of rendered text for
external text, embedded text, 608 and 708 based subtitles. Teletext
subtitles are unchanged as the text does not appear to be cropped and
the positioning of text is less flexible.

Closes #8954


git-svn-id: http://svn.mythtv.org/svn/trunk@26423 7dbf422c-18fa-0310-86e9-fd20926502f2