Fix perl bindings make install
Instead of passing PREFIX in directly, if you want to install the Perl bindings
anywhere but system default, you will need to use --perl-config-opts in 
configure like so:
 --perl-config-opts="PREFIX=/tmp/blah"

Of course substitute in your selected PREFIX value

If you do install them in nonstandard places, it is still up to you as the user
to set @INC or PERL5LIB correctly so your scripts can see the bindings.

Fixes #8973.



git-svn-id: http://svn.mythtv.org/svn/trunk@26427 7dbf422c-18fa-0310-86e9-fd20926502f2