Fix the livetv channel editor.
The problem was that the Data Direct update was offloaded to a separate
thread that then created the Channel Editor OSD dialog.

We now create the dialog immediately and then spawn the update thread
which will re-populate the fields if there is new/different data. Only
the UI thread can now create the dialog.

Closes #8760


git-svn-id: http://svn.mythtv.org/svn/trunk@26432 7dbf422c-18fa-0310-86e9-fd20926502f2