Improved CrystalHD support.
- revert back to passing raw timestamps to the decoder.
- implement a packet buffer to store packets in case the decoder's
buffers are full.
- use a separate thread to retrieve decoded frames from the decoder,
which should avoid deadlocks when the input buffers are full.
- enable support for WMV3 and VC1 streams. I have only one VC1 clip that
doesn't play, which hopefully is a rare exception.
- break out the packet processing from GetFrame which should simplify
subsequent changes needed to drain the decoded video buffers.

Under certain circumstances, playback will now deadlock when the audio
buffers are full and the decoder has not had the opportunity to release
its store of decoded frames. This should be fixed before 0.24.


git-svn-id: http://svn.mythtv.org/svn/trunk@26464 7dbf422c-18fa-0310-86e9-fd20926502f2