Fix seeking to near the end of an in progress recording.
We were hitting an infinite loop when frame by frame seeking as the
position map wasn't actually getting updated - hence we were never going
to break out of the seek loop for a file that didn't have an end (i.e.
in progress).

I'm not sure whether this is a long standing issue that has been exposed
or a subtle side effect of the thread changes in MythPlayer.

Closes #8842 - though I'll also later add some sort of debug check to
identify any other cases that trigger this condition.


git-svn-id: http://svn.mythtv.org/svn/trunk@26486 7dbf422c-18fa-0310-86e9-fd20926502f2