Better handle pause-state tracking when the "Already Editing" dialog, which pauses playback, is displayed.
This fixes a user-experience issue where the frontend seems to lock if the "Already Editing" dialog is displayed for a show without a seek table.  Instead, playback was just left paused, though the user expected playback to continue, as before attempting to edit the recording.  Thanks to Stuart Morgan for noticing the issue.


git-svn-id: http://svn.mythtv.org/svn/trunk@26502 7dbf422c-18fa-0310-86e9-fd20926502f2