Added kMSProcessEvents to myth_system to allow for previous behavior where blocking the UI thread with MYTH_SYSTEM_DONT_BLOCK_PARENT would be the only time that we would process the UI events while we wait for the PID to finish. This allows us to selectively process the events when running in the UI thread.
This should fix the deadlock that markk saw earlier where checking the status
of pulseaudio, if some particular events come in while we are waiting, we may
deadlock.  This was due to the timerEvent in tv_play.cpp using
GetPlayerWriteLock in the keyEvent handling, and the pulseaudio handling runs
inside that lock.  Some events (in particular DONE_RECORDING) have code that
use GetPlayerReadLock, and this causes a deadlock.

For this particular call to myth_system, I have disabled processing of events
to keep this particular deadlock from happening.  There may be other places
that require this treatment as well.



git-svn-id: http://svn.mythtv.org/svn/trunk@26526 7dbf422c-18fa-0310-86e9-fd20926502f2