Remove the audio.IsPaused() check from MythPlayer::Pause().
Sometimes audio playback takes time (up to about 25ms) to finish pausing, so on some systems, IsPaused() would return false when we were in the process of pausing, making the return value of MythPlayer::Pause() incorrect.  This resulted in the player getting into a weird paused-but-playing state, for example, after editing a recording if the player was paused before entering edit mode.  Thanks to Mark Kendall for the suggested fix.


git-svn-id: http://svn.mythtv.org/svn/trunk@26533 7dbf422c-18fa-0310-86e9-fd20926502f2