DVD playback improvements.
- move the still frame timer out of the ringbuffer and into the player.
This allows us to check the still frame status from both the decoder
thread and the playback thread, avoiding deadlocks when the decoder
thread is stalled and hence the still frame timeout check is not
triggered in the ringbuffer.

- refactor MythDVDPlayer::VideoLoop to adjust the timing of certain
events and avoid at least one deadlock obeserved in short (3 second)
video sequences.

Still plenty of other fixes required.


git-svn-id: http://svn.mythtv.org/svn/trunk@26646 7dbf422c-18fa-0310-86e9-fd20926502f2