Fix tab order and focus issues in video sources setup.  Fixes #8726 ( comment:7:ticket:8726 ).  Refs #6637 ( comment:1:ticket:6637 ).
This was broken in [17597].  Because the XMLTV_generic_config widgets were added to the XMLTVConfig after the XMLTVConfig is created and displayed, they were added as (stacked, obscured) visible widgets, and, therefore, receive focus when tabbing.  The patch basically reverts the non-blocking part of [17597].  In the worst case, the blocking may cause a delay of 25s (until timeout), but generally should be much shorter.  The change, however, makes the screen functional and, since the screen is a Qt-based screen whose code will be replaced in the setup rewrite, I don't feel it's worth trying to salvage the non-blocking functionality.


git-svn-id: http://svn.mythtv.org/svn/trunk@26662 7dbf422c-18fa-0310-86e9-fd20926502f2