Switch over to using the new PulseHandler class.
The PulseAudio suspend/resume code now uses a global singleton to
interact with the sound server. It should be entirely thread safe,
handle unexpected callbacks from the server in a graceful manner, be
faster in operation (we no longer connect each time) and it does not
require an additional thread to operate.

A couple of items still need cleaning up:-

- need to add back a check into IsPulseAudioRunning when pulse support
is not compiled in and warn the user when PulseAudio is running (once on
startup)
- really need to cleanup the connection on exit
(Suspend(PulseHandler::kPulseCleanup)) to ensure we don't leak
connections in the server.

Refs #9031


git-svn-id: http://svn.mythtv.org/svn/trunk@26703 7dbf422c-18fa-0310-86e9-fd20926502f2