DVD playback changes.
N.B. This is not complete - there are additional changes to come so
don't expect everything to work just yet.

N.B. Current known issues are broken DVD bookmarking and potential
deadlocks when using XVideo.

- remove the DVD state change handling/parsing from AvFormatDecoder.
Trying to synchronise changes between the decoder, the DVD ringbuffer
object and the player is never going to work satisfactorily and I think
was largely a workaround for other issues - some of which are already
fixed and others that will be fixed.

- add a DVD specific HandleStreamsChanged callback. This does not rescan
the video stream (which caused endless problems) or reset the
videobuffers (which causes obvious discontinuities). There are
additional changes to come here as we need to be smarter about how we
trigger and respond to stream changes.

- add an additional waiting state to DVDRingBuffer that waits for the
player to drain all of its existing buffers before we continue. This
ensures we always start the new cell once the last one has played out,
smooths transitions and should by and large fix missing still frames and
menu buttons.

- rework some of the state handling in MythDVDPlayer to manage
cell/title transitions.


git-svn-id: http://svn.mythtv.org/svn/trunk@26768 7dbf422c-18fa-0310-86e9-fd20926502f2