Modify the access specifiers in MythPlayer and MythDVDPlayer.
This changes a number of methods from public to protected access with a
view to more rigorously restricting the decoder's access to the player
for thread safety reasons. I've also, for the time being, highlighted
the remaining public methods that are dangerous (almost exclusively
driven by the DVD code).

No functional changes.


git-svn-id: http://svn.mythtv.org/svn/trunk@26774 7dbf422c-18fa-0310-86e9-fd20926502f2