DVD initialisation fixes.
- in libdvdnav, title 0 is the menu 'title', so query and report the
number of parts for the actual range of titles on the disc.
- when performing the initial scan, play from the first actual title
(number 1) to avoid failing with discs where the menu has no useful
audio or video data.
- reset the dvdnav object prior to real playback and use the 'default'
dvd_title_play method with title '0' to play from the beginning. This
avoids the custom dvd_first_play method which crashes on some titles
(and is about to be removed).


git-svn-id: http://svn.mythtv.org/svn/trunk@26786 7dbf422c-18fa-0310-86e9-fd20926502f2