More DVD fixes.
- rely entirely on the the DVDNAV_CELL_CHANGED event details to
determine where we are in playback (title, part, still frame and menu)
and use this information to decide when we need to force a resync with
the player.
- call ClearAfterSeek after we've forcibly drained the video buffers to
ensure we restart av sync.
- force a low prebuffering threshold to speed up title/menu transitions.
- some extra/improved debugging.
- fix spurious still frame timeouts that were causing entirely
unexpected behaviour.

In combination these changes fix a number of discs that simply did not
play, various av sync/playback issues and a number of menu items that
either did not work or behaved unusually.


git-svn-id: http://svn.mythtv.org/svn/trunk@26813 7dbf422c-18fa-0310-86e9-fd20926502f2