Use the MPEG sequence end marker to trigger capturing of the DVD still frame.
When we are in a DVD menu that uses a still frame with audio, these are
not flagged by libdvdnav as still frames. We thus need a different
trigger to ensure the still frame is not missed. Generally speaking the
sequence end marker precedes the first frame of the new sequence.

This reinstates, in a slightly different way, the behaviour of the old
DVD code.


git-svn-id: http://svn.mythtv.org/svn/trunk@26838 7dbf422c-18fa-0310-86e9-fd20926502f2