Fix still frame DVD menus with audio.
Instead of using the pause frame code, which was the previous approach
for menus, we simply use the last frame displayed (but without releasing
it for re-use), update (hack) its timecode and continue to run the av
sync code. This ensures the av sync does not get wildly broken, dvd
buttons continue to display and transitions between still menus and full
av menus work seemlessly.


git-svn-id: http://svn.mythtv.org/svn/trunk@26841 7dbf422c-18fa-0310-86e9-fd20926502f2