Fix unexpected MythPlayer exits by clearing the buffering state.
The pre-buffering check in MythPlayer::PreBufferEnoughFrames implicitly
assumes that it is called sequentially during playback and hence any
buffering state is cleared. If playback is paused, or an alternative
playback route used (e.g. DVD playback), then the buffering state may
not be cleared, the timer continues to roll and playback promptly exits
when normal playback is resumed.


git-svn-id: http://svn.mythtv.org/svn/trunk@26887 7dbf422c-18fa-0310-86e9-fd20926502f2