Backport [26889] from trunk.
Fixes the seemingly uncommon occurance of av_read_frame returning -EAGAIN.
When this happens, we are supposed to try again, not assume that the file hit
EOF or errored out. This was hitting me hard on a pathologically bad recording
of 30 Rock. It would just prematurely stop playback or seektable rebuild at
about 23% through the file.



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-24-fixes@26890 7dbf422c-18fa-0310-86e9-fd20926502f2