Fix playback of some DVDs.
When we initially open a DVD and scan it for av streams, we
intentionally ignore certain dvdnav events and continue reading from the
disk. Sometimes this means libdvdnav emits a DVDNAV_STOP event
(equivalent to an EOF), a state which does not seem to be cleared when
we reset the device for actual playback. libdvdnav will not return any
more packets for decoding.

We now close and re-open the disc if a DVDNAV_STOP event has been
emitted during the initial scan, which appears to work in the limited
number of discs that have this problem.


git-svn-id: http://svn.mythtv.org/svn/trunk@26920 7dbf422c-18fa-0310-86e9-fd20926502f2