Backport r26895 from trunk to release-0-24-fixes.
Refactor MythPlayer::AVSync to allow it to use the pause frame as the
displayed frame.

VideoOutput::PrepareFrame and VideoOutput::ProcessFrame will both use
the pause frame if passed a NULL VideoFrame pointer. This is the first
of two commits to address the mistake I made in fixing DVD audio still
frame menus - namely that we have a pause frame for a reason and re-
video frame, which is fine with OpenGL and VDPAU but not with XVideo.


git-svn-id: http://svn.mythtv.org/svn/branches/release-0-24-fixes@26926 7dbf422c-18fa-0310-86e9-fd20926502f2