Backports [26963] from trunk.
Here's the original commit message:

Don't open an event socket when a backend connects back to the
master backend via MythCoreContext::ConnectToMasterServer().
Normally this shouldn't be necessary since slaves should be
using their existing master backend connection, but this fix
will prevent an event loop should something cause a slave to
connect back to the master.

Refs #9123.  This fixes the event loop in #9123, but doesn't
fix the actual bug that is causing the slave to connect back
to the master, so I'm leaving the ticket open for now.  I'm
also lowering the status level of the ticket separately
since #9123 should no longer be a blocker.



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-24-fixes@26964 7dbf422c-18fa-0310-86e9-fd20926502f2