Add a lock around the OSD object in MythPlayer.
With the recent changes in MythPlayer, the OSD is accessed and modified
almost exclusively from the main playback/UI thread. The one exception
to this is when the decoder triggers a video stream change and the
VideoOutput object is recreated - which may recreate the OSD painter.

Post 0.24, I'll move the video object recreation into the main thread as
well but it is a fairly significant change with huge potential for
locking up the player - and hence it needs to wait.


git-svn-id: http://svn.mythtv.org/svn/trunk@27011 7dbf422c-18fa-0310-86e9-fd20926502f2