Backport r26999 from trunk to release-0-24-fixes.
Improved tracking of allocated images in the Painter classes.

    * add a QList to the base painter class to keep a record of all
image objects allocated by the painter and subsequently de-allocated by
the parent UI object.

    * Move painter specific image deletion into private methods and
ensure that DeleteFormatImage is only called by the parent object when
it no longer has a need for the allocated image, with all other internal
deletion using the private method.


git-svn-id: http://svn.mythtv.org/svn/branches/release-0-24-fixes@27027 7dbf422c-18fa-0310-86e9-fd20926502f2