Backport r27014 from trunk to release-0-24-fixes.
Improved MythImage? re-parenting.

We only need to reset a MythImage?'s parent when it has none (the
previous parent painter was deleted) - this fixes an issue where images
allocated by the main UI painter but then painted within the OSD were
reparented to the OSD painter, which caused a crash when the frontend
exited.

If an image is already parented to one painter and we are creating a
cached bitmap in another painter, it will either be expired as needed
when the cache size is exceeded or routinely deleted when the painter is
deleted.


git-svn-id: http://svn.mythtv.org/svn/branches/release-0-24-fixes@27033 7dbf422c-18fa-0310-86e9-fd20926502f2