Backport r27015 from trunk to release-0-24-fixes.
Fix ownership/allocation of MythImages?.

There are a number of instances within the UI code where new images are
allocated by asking MythMainWindow? for its painter. When this code is
used within the OSD, images are then allocated using the incorrect
painter and we rapidly lose track of who owns what.

We now keep a record of the current painter within the UI object tree
itself and search the object tree for the correct painter to use each
time we need a new image. For the main UI, this will still ultimately
return the main UI painter within the screenstack objects and for the
OSD, the painter is set and updated at the ScreenType? (OSD window)
level.

There is still an intermittent crash on mythfrontend exit, related to
the use of an invalid parent within MythImage?, that needs tracking
down.


git-svn-id: http://svn.mythtv.org/svn/branches/release-0-24-fixes@27034 7dbf422c-18fa-0310-86e9-fd20926502f2