Fix the H.264 seektable rebuild.
Turns out that the GetFrame() would return for both top and bottom field (in
1080i) and the mythcommflagplayer would count each field as a frame, even
though, internally, the seektable building had proper *frame* numbers.  The fix
was to propogate up correctly whether or not a new frame was found, and only
count frames.

As the saving to database is done in chunks, after the first chunk is saved,
the mythcommflagplayer's view of where the second chunk started didn't match
reality, so it very happily saved empty chunks of the map to the DB, resulting
in a very short recording length being seen as only the first chunk made it to
the db.

This has been tested against HD-PVR generated H.264, and regression tested with
MPEG2 recordings to make sure I didn't mess them up.

Fixes #9126



git-svn-id: http://svn.mythtv.org/svn/trunk@27098 7dbf422c-18fa-0310-86e9-fd20926502f2